--- packages/domains/encryption/__tests__/encrypt.test.ts
@@ -1,3 +1,0 @@
-test("should test", () => {
-  expect(1).toBe(1);
-});
