+++ packages/domains/index.ts
@@ -1,3 +1,7 @@
-import _exampleDomain from "./handlers/example-handler";
+import _secrets from "./src/secrets";
+import _encryption from "./src/encryption";
 
-export const exampleDomain = _exampleDomain;
+export default {
+  secrets: _secrets,
+  encryption: _encryption
+};
