+++ packages/domains/src/encryption/index.ts
@@ -1,0 +1,9 @@
+import SecretManager from "../secrets";
+
+import _encrypt from "./encrypt";
+import _decrypt from "./decrypt";
+
+export default {
+  encrypt: _encrypt(SecretManager),
+  decrypt: _decrypt(SecretManager),
+};
